Add files via upload
PMFG files are short for "Play My Favorite Games" The programs take a file on my home computer, that containes a game and an associated file path. The game and file path are added to a dictionary (HashMap) and then i can select a game from the dictionary. They can also be added to a regular array and then selected via an associated number. This has been done in powershell and python. Thank You for looking! :)